Improved lose coupling and portability
Moved the File operations into a seprate 'model'-Modul.
Moved the SMTP-connect/disconnect into 'sendmail' function to
decouple the process of mail sending from the program logic in
'SiteCheck' modul.
Diffs are now generated by a Python Libary Modul instead of an
LINUX diff tool to improve portability.

The incorporated Changes are intended to make it possible to port
a version of SiteCheck for Google App Engine.