Major Changes everywhere
Sitecheck now uses an model based on objects that handle Sites,
Logins and Recipients of update mails. More Steps in the site
checking process are placed in own functions. The datastorge is
done in a seprate 'modle.py' module. Minor changes were made to
'sendmail.py' to work properly in the new system. 'sites.py' has
become usesless and was deleted.